+++ changelog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Bugfix: Fixed parsing liferay-look-and-feel.xml when creating that file for the first time
 - Bugfix: Correctly detect locale strings in JSON schema for journal structure files
 - Bugfix: Fixed compatibility with IntelliJ 2018.3 when injecting javascript into freemarker taglibs
+- Feature: Code completion for portlet names in JSP taglibs
 
 ## 0.0.3
 - Feature: Javascript injection for aui:validator and many onXXX attributes